--- SPM_API.py
@@ -1,4 +1,0 @@
-# Simple Password Manager Application Programming Interface
-
-# Import SPM Python packages
-from SPM import *
